add column active and updating endpoints

+++ crud/src/main/java/com/kethlyn/crud/domain/product/RequestProduct.java
@@ -2,6 +2,8 @@ package com.kethlyn.crud.domain.product;
 
 import jakarta.validation.constraints.NotEmpty;
 
-public record RequestProduct(String id,String name,Integer price) {
+import java.math.BigDecimal;
+
+public record RequestProduct(String id, String name, BigDecimal price) {
 
 }
