+++ crud/src/main/resources/db/migration/V1__create-table-product.sql
@@ -1,5 +1,5 @@
 CREATE TABLE product(
     id TEXT PRIMARY KEY UNIQUE NOT NULL,
     name TEXT NOT NULL,
-    price INT NOT NULL
+    price DECIMAL(10,2) NOT NULL
 );
